Update lint settings and fix issues.

+++ skeleton/BOOT/miyoo/app/my355.sh
@@ -18,7 +18,7 @@ export LD_LIBRARY_PATH=/usr/miyoo/lib:/usr/lib:/lib
 
 # .tmp_update/updater does the actual installation (and later, updating)
 cp -rf .tmp_update $SDCARD_PATH/
-rm -rf "$SDCARD_PATH/$MIYOO_DIR"
+rm -rf "${SDCARD_PATH:?}/${MIYOO_DIR:?}"
 sync
 $SDCARD_PATH/.tmp_update/updater
